refactor: move logging and error message logic
