+++ src/loggin-error-message.ts
@@ -1,0 +1,50 @@
+import Color from "../color-text/Color";
+import { lineBreak } from "../color-text/styling-terminal";
+import { TestCaseError } from "./TestsPerSpecFile";
+import { log } from "./loggin-results";
+import { styleFileLocation } from "./test-results";
+
+/**
+ * Highlight the error indicator in the code snippet
+ * @param codeSnippet
+ */
+export const highlightErrorIndicator = (codeSnippet: string): string => {
+    const lines = codeSnippet.split("\n");
+    const styledLines = lines.map((line) => {
+        if (line.includes("^")) {
+            return Color.text(line).red().valueOf();
+        } else {
+            return line;
+        }
+    });
+
+    return styledLines.join("\n");
+};
+
+
+/**
+ * Log the test error message,
+ * with a code snippet and a file location link
+ * @param codeSnippet
+ */
+export const logTestError = (failedTests: TestCaseError[]) => {
+    let counter = 0;
+    failedTests.forEach((failedTest) => {
+        const error = failedTest.error;
+        //TODO fix me - issue on the log of the file location - it is not correct
+        const titlePath = failedTest.testData.title;
+        const title = `${titlePath[2]} > ${titlePath[3]} > ${titlePath[4]} ───────────────`;
+        log(Color.text(`${++counter}) ${title}`).red().valueOf());
+        log(`${error.message}`);
+        if (error.value !== undefined) {
+            log(error.value);
+        }
+        if(error.snippet !== undefined) {
+            const highlightedError = highlightErrorIndicator(error.snippet);
+            log(highlightedError);
+        }
+        const fileLocationStyle = styleFileLocation(error.location);
+        log(`\t${Color.text('at').gray().valueOf()} ${fileLocationStyle}`);
+        log(lineBreak);
+    });
+};
